+++ .github/CHANGELOG.md
@@ -17,6 +17,7 @@
 1. [HYD] Added hydraulic reverser actuators - @Crocket63 (crocket)
 1. [EIS] Added > character to the font - @KiloEchoVictor (Kevin)
 1. [MCDU] The remote MCDU protocol now sends the annunciator light state - @tracernz (Mike)
+1. [MCDU] The remote MCDU protocol now sends the display and integral light brightness - @tracernz (Mike)
 
 ## 0.10.0